Status report:
The current work-in-progress is to avoid requeueing links which have
already been seen. Because there is one goroutine running for each page
that we wish to scrape we have a separate goroutine processing each of
the links we might wish to follow, which manages our central state of
what we've seen.

Prior to this everything was working delightfully. The tasks WaitGroup
unlocked when it was supposed to and the program exited after completing
all of the work it intended to.

Now, however, the program is exiting too early. Debugging...

... to resume after breakfast.